+++ src/multi_vector_simulator/F1_plotting.py
@@ -24,6 +24,7 @@ from multi_vector_simulator.utils.constants import (
     ECONOMIC_DATA,
     LABEL,
     OUTPUT_FOLDER,
+    SOC,
 )
 
 from multi_vector_simulator.utils.constants_json_strings import (
@@ -849,9 +850,9 @@ def create_plotly_flow_fig(
 
     assets_list = list(df_plots_data.columns)
     assets_list.remove("timestamp")
-    if any("SOC" in item for item in assets_list):
+    if any(SOC in item for item in assets_list):
         # remove SOC as it is provided in % and does not fit to this graph
-        assets_list = [s for s in assets_list if "SOC" not in s]
+        assets_list = [s for s in assets_list if SOC not in s]
         # call SOC plot here todo
 
     for i, asset in enumerate(assets_list):